test(api): moved UserPreferences tests in User to separate file
